+++ include/study.h
@@ -1,7 +1,8 @@
 #include <flashcards.h>
 
 
-bool get_settings(FlashcardSet *flashcard_set, bool* starred_only, bool* shuffle, bool vectorsin[], bool vectorsout[]);
+bool get_settings(FlashcardSet *flashcard_set, bool* starred_only, bool* shuffle, int* question_count, bool vectorsin[], bool vectorsout[]);
+void editVectors(bool *vectorin,bool *vectorout,int vectorCount);
 void pickMode(char* list);
 
 int getOrder(FlashcardSet *flashcard_set, int order[], bool shuffle, bool star_only);