+++ src/main.c
@@ -48,9 +48,16 @@ int main_menu_keybinds(void* menu);
 
 int main(int argc, char *argv[]){
 
-    if(argc>1){
-        if(strcmp(argv[1],"--version")==0){
-            printf("TUIFlashcards version %lf", ReleaseVersion);
+    for(int i = 1; i < argc; i++){
+        if(strcmp(argv[i],"--version")==0 || strcmp(argv[i],"-v")==0){
+            printf("TUIFlashcards version %s\n", ReleaseVersion);
+            return 0;
+        }
+        if(strcmp(argv[i],"--help")==0 || strcmp(argv[i],"-h")==0){
+            printf("Usage: \n\t%s [flags]\n\n", argv[0]);
+            printf("Flags:\n");
+            printf("\t--version, -v\t\tPrint version information\n");
+            printf("\t--help, -h   \t\tPrint this help screen\n");
             return 0;
         }
     }